Implemented self-adjusting clock speed
This is the best I can do without a video frame or audio bitrate to tie
execution speed to. It does stabilise at the provided speed, and it
could be improved by making the clock granularity dynamically
adjustable, but just wanted to keep things simple.

The cycle rate will be spammed every second, but the following changes
in the code will focus on implementing a debugger and capturing user
input through special keys, and one of them will trigger the rate to be
printed, for the time being it's left as is just to debug